+++ CSGO/CSGO/Source/dllmain.h
@@ -26,8 +26,8 @@
 #include "Features/Miscellaneous/automation.h"
 #include "Features/Miscellaneous/fov.h"
 
-void Feature( bool, unsigned, std::function< void( ) > );
-void Feature( bool, unsigned, std::function< void( ) >, int );
+void Feature( bool, unsigned long, std::function< void( ) > );
+void Feature( bool, unsigned long, std::function< void( ) >, unsigned short );
 void CleanUp( );
 void Panic( );
 void CreateThreads( );
@@ -46,9 +46,14 @@ BOOL WINAPI DllMain( HINSTANCE hInstDll, DWORD fdwReason, LPVOID lpvReserved );
  * * * typedef
  * * * b bool/byte
  * * * f FLAG
+ * * * n nano (__int8)
+ * * * un unsigned nano (unsigned __int8)
+ * * * s short int
+ * * * us unsigned short
  * * * i int
- * * * n unsigned int (total, c are alternatives - pls do not use these)
+ * * * u unsigned int (total, c are alternatives - pls do not use these)
  * * * l long
+ * * * ul unsigned long
  * * * clr color
  * * * cor coordinate
  * * * ang angle
